+++ modules/darwin/apps/default.nix
@@ -21,6 +21,7 @@ _:
     ./github-runner-container.nix
     ./openbao-aws-creds.nix
     ./openbao-github-creds.nix
+    ./openbao-run.nix
     ./orbstack.nix
     ./raycast.nix
     ./screen-sharing.nix